feat: create service to search movie by name
In movie service the function was created with endpoint to search just movies by name.
In Navigation the button to search handles submit the term entered in the input.
In Navigation the buttons to switch to movies or series also clear the term from the search and call handlesubmit function to update that, so that the state of search is reset.
In home the function to handle search is created and passed through props to Navigate. Also, a ternary had to be added to condition the renderization of the list of movies, depending on whether there is a search or not.
Some styles were adjusted in home, login, and index.